requirements.txt
streamlit==1.41.1
google-generativeai==0.8.4
notion-client==2.2.1
pandas==2.2.3
Pillow==11.1.0
python-dotenv==1.0.1
